requirements.txt
argon2-cffi==16.3.0
arrow==0.10.0
awsebcli==3.10.4
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.4.7
botocore==1.7.11
cement==2.8.2
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
colorama==0.3.7
cycler==0.10.0
decorator==4.1.1
Django==1.11.3
django-crispy-forms==1.6.1
django-forms-bootstrap==3.1.0
django-html5==1.0.0
django-mysql==2.1.0
django-s3direct==1.0.3
django-storages==1.6.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
lxml==3.8.0
matplotlib==2.0.2
mongo==0.2.0
mysql-connector-python==2.1.6
mysql-connector-python-rf==2.0.4
MySQL-python==1.2.5
mysqlclient==1.3.4
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
pandas-datareader==0.4.0
pathspec==0.5.0
Pillow==4.2.1
public==0.0.38
pycparser==2.18
pygame==1.9.1
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-http-client==2.4.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
self==0.0.14
semantic-version==2.5.0
setupfiles==0.0.50
six==1.10.0
stripe==1.67.0
tabulate==0.7.5
termcolor==1.1.0
tweepy==3.5.0
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
whitenoise==4.0b2
